{"schema":"decodex/commit/1","summary":"Split ELF CLI benchmark argument models into a private child module after local validation.","authority":"manual"}
